--- a/Microsoft.Azure.Cosmos/src/CosmosElements/CosmosArray.LazyCosmosArray.cs
+++ b/Microsoft.Azure.Cosmos/src/CosmosElements/CosmosArray.LazyCosmosArray.cs
@@ -3,8 +3,9 @@
 //------------------------------------------------------------
 namespace Microsoft.Azure.Cosmos.CosmosElements
 {
+#nullable enable
+
     using System;
-    using System.Collections;
     using System.Collections.Generic;
     using System.Linq;
     using Microsoft.Azure.Cosmos.Json;
@@ -20,10 +21,9 @@ namespace Microsoft.Azure.Cosmos.CosmosElements
     {
         private sealed class LazyCosmosArray : CosmosArray
         {
+            private readonly Lazy<List<CosmosElement>> lazyCosmosElementArray;
             private readonly IJsonNavigator jsonNavigator;
             private readonly IJsonNavigatorNode jsonNavigatorNode;
-            private readonly Dictionary<int, CosmosElement> cachedCosmosElements;
-            private readonly Lazy<int> lazyCount;
 
             public LazyCosmosArray(
                 IJsonNavigator jsonNavigator,
@@ -37,81 +37,29 @@ namespace Microsoft.Azure.Cosmos.CosmosElements
 
                 this.jsonNavigator = jsonNavigator;
                 this.jsonNavigatorNode = jsonNavigatorNode;
-                this.cachedCosmosElements = new Dictionary<int, CosmosElement>(capacity: 0);
-                this.lazyCount = new Lazy<int>(() => this.jsonNavigator.GetArrayItemCount(this.jsonNavigatorNode));
-            }
 
-            public override int Count => this.lazyCount.Value;
-
-            public override CosmosElement this[int index]
-            {
-                get
+                this.lazyCosmosElementArray = new Lazy<List<CosmosElement>>(() =>
                 {
-                    if (!this.cachedCosmosElements.TryGetValue(index, out CosmosElement cachedValue))
+                    List<CosmosElement> elements = new List<CosmosElement>(capacity: jsonNavigator.GetArrayItemCount(jsonNavigatorNode));
+                    // Using foreach instead of indexer, since the navigator doesn't support random seeks efficiently.
+                    foreach (IJsonNavigatorNode arrayItem in jsonNavigator.GetArrayItems(jsonNavigatorNode))
                     {
-                        IJsonNavigatorNode jsonNavigatorNode = this.jsonNavigator.GetArrayItemAt(this.jsonNavigatorNode, index);
-                        cachedValue = CosmosElement.Dispatch(this.jsonNavigator, jsonNavigatorNode);
-                        this.cachedCosmosElements[index] = cachedValue;
+                        elements.Add(CosmosElement.Dispatch(jsonNavigator, arrayItem));
                     }
 
-                    return cachedValue;
-                }
+                    return elements;
+                });
             }
 
-            public override IEnumerator<CosmosElement> GetEnumerator() => new LazyCosmosArrayEnumerator(this);
+            public override int Count => this.lazyCosmosElementArray.Value.Count;
+
+            public override CosmosElement this[int index] => this.lazyCosmosElementArray.Value[index];
+
+            public override IEnumerator<CosmosElement> GetEnumerator() => this.lazyCosmosElementArray.Value.GetEnumerator();
 
             public override void WriteTo(IJsonWriter jsonWriter) => this.jsonNavigator.WriteNode(this.jsonNavigatorNode, jsonWriter);
 
             public override IJsonReader CreateReader() => this.jsonNavigator.CreateReader(this.jsonNavigatorNode);
-
-            private struct LazyCosmosArrayEnumerator : IEnumerator<CosmosElement>
-            {
-                private readonly LazyCosmosArray lazyCosmosArray;
-                private IEnumerator<IJsonNavigatorNode> nodes;
-                private int index;
-
-                public LazyCosmosArrayEnumerator(LazyCosmosArray lazyCosmosArray)
-                {
-                    this.lazyCosmosArray = lazyCosmosArray;
-                    this.Current = default;
-                    this.nodes = default;
-                    this.index = default;
-                    this.Reset();
-                }
-
-                public CosmosElement Current { get; private set; }
-
-                object IEnumerator.Current => this.Current;
-
-                public void Dispose()
-                {
-                }
-
-                public bool MoveNext()
-                {
-                    if (!this.nodes.MoveNext())
-                    {
-                        this.Current = default;
-                        return false;
-                    }
-
-                    this.index++;
-                    if (!this.lazyCosmosArray.cachedCosmosElements.TryGetValue(this.index, out CosmosElement cachedValue))
-                    {
-                        cachedValue = CosmosElement.Dispatch(this.lazyCosmosArray.jsonNavigator, this.nodes.Current);
-                        this.lazyCosmosArray.cachedCosmosElements[this.index] = cachedValue;
-                    }
-
-                    this.Current = cachedValue;
-                    return true;
-                }
-
-                public void Reset()
-                {
-                    this.nodes = this.lazyCosmosArray.jsonNavigator.GetArrayItems(this.lazyCosmosArray.jsonNavigatorNode).GetEnumerator();
-                    this.index = -1;
-                }
-            }
         }
     }
 #if INTERNAL