… Microsoft.Azure.Cosmos/src/CosmosElements/CosmosObject.cs
… Microsoft.Azure.Cosmos/src/Query/Core/Pipeline/Aggregate/Aggregators/SingleGroupAggregator.cs
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.EmulatorTests/Query/OrderByQueryTests.cs
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Performance.Tests/CosmosElements/LazyCosmosElementsBenchmark.cs
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Performance.Tests/CosmosElements/LazyCosmosElementsDeserializerBenchmark.Results.txt
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/ChangeFeed/ChangeFeedEstimatorIteratorTests.cs
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/CosmosElements/EquatableTests.cs
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/CosmosElements/LazyCosmosElementTests.cs
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/Pagination/DocumentContainerTests.cs